+++ include/BField/DipoleB.h
@@ -9,14 +9,14 @@ class DipoleB : public BModel
 {
 protected:
 	//Field simulation constants
-	meters  L_m{ 0.0 };
-	meters  L_norm_m{ 0.0 };
-	meters  s_max_m{ 0.0 };
+	meters  L_m{ 0.0f };
+	meters  L_norm_m{ 0.0f };
+	meters  s_max_m{ 0.0f };
 
 	//specified variables
-	degrees ILAT_m{ 0.0 };
-	meters  ds_m{ 0.0 };
-	ratio   lambdaErrorTolerance_m{ 0.0 };
+	degrees ILAT_m{ 0.0f };
+	meters  ds_m{ 0.0f };
+	ratio   lambdaErrorTolerance_m{ 0.0f };
 
 	bool    useGPU_m{ true };
 
@@ -37,13 +37,13 @@ public:
 	__host__            degrees ILAT()  const;
 
 	__host__ __device__ tesla   getBFieldAtS(const meters s, const seconds t) const override;
-	__host__ __device__ double  getGradBAtS (const meters s, const seconds t) const override;
+	__host__ __device__ float  getGradBAtS (const meters s, const seconds t) const override;
 	__host__ __device__ meters  getSAtAlt   (const meters alt_fromRe) const override;
 
 	__host__            ratio  getErrTol() const;
 	__host__            meters getds()     const;
 
-	__host__            vector<double> getAllAttributes() const override;
+	__host__            vector<float> getAllAttributes() const override;
 	__host__            void serialize(ofstream& out) const override;
 };
