+++ include/utils/numerical.h
@@ -16,14 +16,14 @@ namespace utils
 {
 	namespace numerical
 	{
-		DLLEXP void v2DtoEPitch(const vector<double>& vpara, const vector<double>& vperp, double mass, vector<eV>& energies, vector<degrees>& pitches);
-		DLLEXP void EPitchTov2D(const vector<eV>& energies, const vector<degrees>& pitches, double mass, vector<double>& vpara, vector<double>& vperp);
-		DLLEXP vector<double> generateSpacedValues(double start, double end, int number, bool logSpaced, bool endInclusive);
-		DLLEXP void normalize(vector<double>& normalizeMe, double normFactor, bool inverse = false);
-		DLLEXP double calcMean(const vector<double>& calcMyMean, bool absValue = false);
-		DLLEXP double calcStdDev(const vector<double>& calcMyStdDev);
-		DLLEXP void coutMinMaxErr(const vector<double>& basevals, const vector<double>& testvals, string label="", bool skipzeroes = true);
-		DLLEXP void coutNumAboveErrEps(const vector<double>& basevals, const vector<double>& testvals, double errEps, string label="", bool skipzeroes = true);
+		DLLEXP void v2DtoEPitch(const vector<float>& vpara, const vector<float>& vperp, float mass, vector<eV>& energies, vector<degrees>& pitches);
+		DLLEXP void EPitchTov2D(const vector<eV>& energies, const vector<degrees>& pitches, float mass, vector<float>& vpara, vector<float>& vperp);
+		DLLEXP vector<float> generateSpacedValues(float start, float end, int number, bool logSpaced, bool endInclusive);
+		DLLEXP void normalize(vector<float>& normalizeMe, float normFactor, bool inverse = false);
+		DLLEXP float calcMean(const vector<float>& calcMyMean, bool absValue = false);
+		DLLEXP float calcStdDev(const vector<float>& calcMyStdDev);
+		DLLEXP void coutMinMaxErr(const vector<float>& basevals, const vector<float>& testvals, string label="", bool skipzeroes = true);
+		DLLEXP void coutNumAboveErrEps(const vector<float>& basevals, const vector<float>& testvals, float errEps, string label="", bool skipzeroes = true);
 	}
 }
